refactor(dashboard): extract QuickViewWorkoutModal from DashboardModals
Bloco inline de 158 linhas do quickViewWorkout (o maior single-block do arquivo) extraído para src/app/(app)/dashboard/modals/QuickViewWorkoutModal.tsx — componente puro, apenas JSX, recebe 3 props (quickViewWorkout, setQuickViewWorkout, handleStartSession).

Mudança mecânica, sem alteração de comportamento:
- DashboardModals.tsx: 759 → 607 linhas (-152)
- 5 ícones lucide agora movidos para o novo arquivo (ArrowLeft, Clock, Dumbbell, Zap, ChevronRight, PlayCircle)
- typecheck ok, eslint ok, 1248/1248 tests passam

Primeiro passo do plano de refactor dos 3 componentes gigantes — começando pelo DashboardModals. WorkoutReport e TeacherManualTab virão em PRs separados pra facilitar revisão visual.
